#108: Re-implement the measure component to work in a multi map scenario. We introduce a MeasureContext class that handles all the layer/interaction management for a particular map.
